OBG-124 fix test
